\docType{data}
\name{palay}
\alias{palay}
\title{Volume of Palay (Rice) Production 1994 - 2013 Quarterly}
\format{A data frame with 79 rows and 7 columns}
\source{
\href{http://countrystat.bas.gov.ph/}{Philippine Bureau of
Agricultural Statistics}
}
\description{
A dataset containing the volume of palay production in
Cordillera Administrative Region (CAR) of the Philippines,
from 1994 - 2013 quarterly.
}
\details{
\itemize{ \item Year. the time variable in quarterly.
\item Apayao. locations 1.  \item Abra. location 2.  \item
Kalinga. location 3.  \item Mt.Province. location 4.  \item
Ifugao. location 5.  \item Benguet. location 6. }
}
\keyword{dataset}
